Articles page with minimal styling and basic functionality
Grabs articles.json, turns the result into an array.
Displays the results in a list, filtered by search terms.
Clicking an article displays the whole markdown file, currently with no formatting at all.

May want to make changes to markdown-to-json so that it supports outputting as an array from the get-go.